+++ commonjs/documentation/changelog.md
@@ -1,6 +1,7 @@
 # Change Log
 <pre>
 v2.2.0  Added Friends, Messages, Events, Files namespaces and methods [MOD-772]
+        Fixed issue with empty dictionary items in data when using OAuth [MOD-817]
 
 v2.1.1  Fixed display issue with custom objects in example application [MOD-770]
         Add check for valid numeric data for key-value increment call in example application [MOD-682]